# src/backend/doc/swagger.yaml
swagger: "2.0"
info:
  description: "TestBuDDy is a test management, requirement management, automation friendly tool."
  version: "1.0.0"
  title: "TestBuDDy"
  contact:
    email: "barbora.blozonova@gmail.com"
  license:
    name: "MIT licence"
host: "127.0.0.1:5000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /projects/{proj_id}/plans/{tp_id}/mods/{tm_id}/cases/{tc_id}:
    get:
      tags:
      - "Test library - cases and steps"
      summary: "Get detailed test case (by set path ID)"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "tp_id"
        description: "test plan id"
        in: "path"
        type: "integer"
        required: true
      - name: "tm_id"
        description: "test module id"
        in: "path"
        type: "integer"
        required: true
      - name: "tc_id"
        description: "test case id"
        in: "path"
        type: "integer"
        required: true
      responses:
        404:
          description: "Project, test plan, test module, or test case not found."
        200:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                type: "array"
                items:
                  allOf:
                    - $ref: "#/definitions/TestCase"
                    - properties:
                        users_assigned:
                          description: "Users assigned to this test case."
                          type: "array"
                          items:
                            type: "integer"
                        scenario_content:
                          type: "string"
          description: "Operation successful. Returns detailed test case in JSON response."
    post:
      tags:
      - "Test library - cases and steps"
      summary: "Add/modify scenario with test steps to test case"
      description: ""
      consumes:
        - "text/plain"
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "tp_id"
        description: "test plan id"
        in: "path"
        type: "integer"
        required: true
      - name: "tm_id"
        description: "test module id"
        in: "path"
        type: "integer"
        required: true
      - name: "tc_id"
        description: "test case id"
        in: "path"
        type: "integer"
        required: true
      - name: "BDD scenario with steps"
        in: "body"
        description: "Scenario with basic steps in JBehave syntax. [text/plain]"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Operation successful. Returns modified test case in JSON response."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                type: "array"
                items:
                  allOf:
                    - $ref: "#/definitions/TestCase"
                    - properties:
                        users_assigned:
                          description: "Users assigned to this test case."
                          type: "array"
                          items:
                            type: "integer"
        400:
          description: "Invalid body or operation failed due to TestBuDDy setting problem."
        404:
          description: "Project, test plan, or test module not found."
        415:
          description: "Unsupported media type."
        500:
          description: " Operation failed."
    put:
      tags:
      - "Test library - cases and steps"
      summary: "Update test case info"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "tp_id"
        description: "test plan id"
        in: "path"
        type: "integer"
        required: true
      - name: "tm_id"
        description: "test module id"
        in: "path"
        type: "integer"
        required: true
      - name: "tc_id"
        description: "test case id"
        in: "path"
        type: "integer"
        required: true
      - name: "test case info"
        in: "body"
        description: "Test case information"
        schema:
          type: "object"
          properties:
            name:
              type: "string"
            priority:
              enum: ["high", "medium", "low"]
              description: "Default priority types: low/medium/high."
              type: "string"
            description:
              type: "string"
            users_assigned:
              description: "Adding users having this test case assigned."
              type: "array"
              items:
                type: "integer"
      responses:
        200:
          description: "Operation successful. Returns test case in JSON response. Affects repository."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                type: "array"
                items:
                  allOf:
                    - $ref: "#/definitions/TestCase"
                    - properties:
                        users_assigned:
                          description: "Users having this test case assigned."
                          type: "array"
                          items:
                            type: "integer"
        400:
          description: "Invalid body."
        404:
          description: "Project, test plan, or test module not found."
        415:
          description: "Unsupported media type."
        500:
          description: " Operation failed."
    delete:
      tags:
      - "Test library - cases and steps"
      summary: "Delete test case"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "tp_id"
        description: "test plan id"
        in: "path"
        type: "integer"
        required: true
      - name: "tm_id"
        description: "test module id"
        in: "path"
        type: "integer"
        required: true
      - name: "tc_id"
        description: "test case id"
        in: "path"
        type: "integer"
        required: true
      responses:
        200:
          description: "Operation successful."
        404:
          description: "Project, test plan, test module or test case not found."
        500:
          description: "Operation failed."
...
definitions:
...
  TestStepDefinition:
    type: "object"
    required:
    - id
    - content
    - project
    - gherkin_type
    properties:
      id:
        type: "integer"
      content:
        type: "string"
      gherkin_type:
        type: "string"
        description: "expected name of step gherkin type"
      param_count:
        type: "integer"
      project:
        type: "object"
        properties:
          id:
            type: "integer"
          name:
            type: "string"
  TestStep:
    type: "object"
    required:
    - id
    - content
    - order
    - test_step_definition
    properties:
      id:
        type: "integer"
      content:
        type: "string"
      order:
        type: "integer"
      test_step_definition:
        $ref: "#/definitions/TestStepDefinition"
  TestCase:
    type: "object"
    required:
    - id
    - name
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      priority:
        enum: ["high", "medium", "low"]
        description: "Default priority types, connected to Test Case."
        type: "string"
      scenario_name:
        type: "string"
        description: "also 'scenario_no_spaces' is used for filenames based on test case"
      steps:
        type: "array"
        items:
          $ref: "#/definitions/TestStep"
...